requirements.txt
requests
discord
beautifulsoup4
python-dotenv